Moved FFmpeg config models to a Config folder.
