+++ Tricycle.Media.FFmpeg/Models/Config/AudioCodec.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-namespace Tricycle.Media.FFmpeg.Models
+namespace Tricycle.Media.FFmpeg.Models.Config
 {
     public class AudioCodec
     {
