+++ Tricycle.UI.Tests/ViewModels/ConfigViewModelTests.cs
@@ -5,16 +5,16 @@ using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using Tricycle.IO;
-using Tricycle.Media.FFmpeg.Models;
+using Tricycle.Media.FFmpeg.Models.Config;
 using Tricycle.Models;
 using Tricycle.Models.Config;
 using Tricycle.UI.Models;
 using Tricycle.UI.ViewModels;
 using Tricycle.Utilities;
-using FFmpegAudioCodec = Tricycle.Media.FFmpeg.Models.AudioCodec;
-using FFmpegAudioConfig = Tricycle.Media.FFmpeg.Models.AudioConfig;
-using FFmpegVideoCodec = Tricycle.Media.FFmpeg.Models.VideoCodec;
-using FFmpegVideoConfig = Tricycle.Media.FFmpeg.Models.VideoConfig;
+using FFmpegAudioCodec = Tricycle.Media.FFmpeg.Models.Config.AudioCodec;
+using FFmpegAudioConfig = Tricycle.Media.FFmpeg.Models.Config.AudioConfig;
+using FFmpegVideoCodec = Tricycle.Media.FFmpeg.Models.Config.VideoCodec;
+using FFmpegVideoConfig = Tricycle.Media.FFmpeg.Models.Config.VideoConfig;
 using TricycleAudioCodec = Tricycle.Models.Config.AudioCodec;
 using TricycleAudioConfig = Tricycle.Models.Config.AudioConfig;
 using TricycleVideoCodec = Tricycle.Models.Config.VideoCodec;
